Extract person import form into a LiveComponent
